Ajustes gerais e bug fixes

+++ src/java/vetorlog/controller/ExpectedPerformanceController.java
@@ -3,13 +3,9 @@ package vetorlog.controller;
 import org.jvnet.hk2.annotations.Service;
 import vetorlog.controller.prototype.Controller;
 import vetorlog.dto.ExpectedPerformanceDTO;
-import vetorlog.dto.GroupSiteDTO;
 import vetorlog.model.ExpectedPerformanceModel;
-import vetorlog.model.GroupSitesModel;
-import vetorlog.model.adapters.ExpectedPerformanceAdapter;
-import vetorlog.model.adapters.TypologyPumpAdapter;
+import vetorlog.model.adapter.ExpectedPerformanceAdapter;
 import vetorlog.serializer.ExpectedPerformanceSerializer;
-import vetorlog.serializer.GroupSiteSerializer;
 
 import javax.inject.Inject;
 import javax.ws.rs.core.Response;
@@ -23,7 +19,7 @@ public class ExpectedPerformanceController extends Controller {
     @Inject
     private ExpectedPerformanceSerializer serializer;
 
-    public Response get(int typologyPumpId, int page, int size) {
+    public Response get(long typologyPumpId, int page, int size) {
         List<ExpectedPerformanceModel> models = dbAdapter.filterByTypologyPumpId(typologyPumpId, page, size);
         List<ExpectedPerformanceDTO> data = serializer.fromModelListToDTOList(models);
         return response.ok(data);
