+++ src/java/vetorlog/controller/TypologyPumpController.java
@@ -4,7 +4,7 @@ import org.jvnet.hk2.annotations.Service;
 import vetorlog.controller.prototype.Controller;
 import vetorlog.dto.TypologyPumpDTO;
 import vetorlog.model.TypologyPumpModel;
-import vetorlog.model.adapters.TypologyPumpAdapter;
+import vetorlog.model.adapter.TypologyPumpAdapter;
 import vetorlog.serializer.TypologyPumpSerializer;
 
 import javax.inject.Inject;
@@ -19,7 +19,7 @@ public class TypologyPumpController extends Controller {
     @Inject
     private TypologyPumpSerializer serializer;
 
-    public Response get(int groupSiteId, int page, int size) {
+    public Response get(long groupSiteId, int page, int size) {
         List<TypologyPumpModel> models = dbAdapter.filterByGroupSiteId(groupSiteId, page, size);
         List<TypologyPumpDTO> data = serializer.fromModelListToDTOList(models);
         return response.ok(data);
